Fix issue in which migrations might be run in the wrong order.
GHC's `getDirectoryContents` is built on top of POSIX readdir(3),
which may or may not (depending on your OS) return entries in sorted
order. On my macOS 10.12 box, it does, whereas on my Ubuntu 16.04 box
they are returned in an (apparently) random order. Since
rivet-autoimporter expects the entries in the `migrations` directory
to be lexically sorted, this was failing on my Ubuntu install but
worked fine on OS X. The fix is to apply `sort` to the result.